Исправление замечаний по коду
Переписывание названий констант прописными буквами.
Перенос функции генерации моков в файл "data.js".
