+++ js/util.js
@@ -1,18 +1,3 @@
-import {
-  Coords,
-  titles,
-  Prices,
-  accommodationTypes,
-  RoomsQuantity,
-  QuestsQuantity,
-  checkinTimes,
-  checkoutTimes,
-  features,
-  descriptions,
-  photos
-} from './data.js';
-
-
 /**
  * Получение случайного числа из диапазона
  * @param {number} min — минимальное значение
@@ -81,61 +66,10 @@ const getRandomQuantityArrayElements = (array) => {
   return shuffleArray(array).slice(0, getRandomNumberInRange(1, array.length));
 }
 
-
-/**
- * Заполнение объявления пользователя мок-данными
- * @return {object} — итоговый объект-объявление
- */
-const renderCard = (i) => {
-  const COORD_X = getRandomNumberInRange(Coords.MIN_X, Coords.MAX_X, Coords.DIGITS);
-  const COORD_Y = getRandomNumberInRange(Coords.MIN_Y, Coords.MAX_Y, Coords.DIGITS);
-
-  const card = {
-    author: {
-      avatar: 'img/avatars/user0' + i + '.png',
-    },
-    offer: {
-      title: getRandomArrayElement(titles),
-      address: COORD_X + ', ' + COORD_Y,
-      price: getRandomNumberInRange(Prices.MIN, Prices.MAX),
-      type: getRandomObjectValue(accommodationTypes),
-      rooms: getRandomNumberInRange(RoomsQuantity.MIN, RoomsQuantity.MAX),
-      guests: getRandomNumberInRange(QuestsQuantity.MIN, QuestsQuantity.MAX),
-      checkin: getRandomArrayElement(checkinTimes),
-      checkout: getRandomArrayElement(checkoutTimes),
-      features: getRandomQuantityArrayElements(features),
-      description: getRandomArrayElement(descriptions),
-      photos: getRandomQuantityArrayElements(photos),
-    },
-    location: {
-      x: COORD_X,
-      y: COORD_Y,
-    },
-  };
-
-  return card;
-};
-
-
-/**
- * Создание массива из карточек-объявлений пользователя
- * @param {number} quantity — количество карточек-объявлений
- * @return {array} — итоговый массив из карточек
- */
-const createCardsList = (quantity) => {
-  let cardsList = [];
-  for (let i = 0; i < quantity; i++) {
-    cardsList.push(renderCard(i));
-  }
-  return cardsList;
-}
-
-
 export {
   getRandomNumberInRange,
   shuffleArray,
   getRandomArrayElement,
   getRandomObjectValue,
-  getRandomQuantityArrayElements,
-  createCardsList
+  getRandomQuantityArrayElements
 };
